+++ src/zernpy/plotting/__init__.py
@@ -1,1 +1,8 @@
+"""
+Default exports from this module (plotting).
+
+@author: @sklykov
+
+"""
+
 __all__ = ['plot_zerns']