+++ src/zernpy/tests/__init__.py
@@ -5,4 +5,4 @@ In this folder, there are tests belonging to the application.
 @author: Sergei Klykov
 @licence: MIT
 """
-__all__ = ['test_calculations', 'test_polynomials_initialization']
+__all__ = ['test_calculations', 'test_polynomials_initialization', 'test_fitting', 'test_calc_psf']