+++ src/zernpy/tests/test_calc_psf.py
@@ -32,49 +32,81 @@ def test_psf_kernel_calc():
     airy_int = get_psf_kernel(zernike_pol=piston, len2pixels=pixel_size, alpha=ampl, wavelength=wavelength, NA=NA,
                               normalize_values=True, airy_pattern=False)
     diff_airy = np.abs(airy_int - airy_p)
-    assert np.max(diff_airy) < 0.01, print("Difference between exact Airy pattern and calculated by the numerical integral is bigger than 1%:",
-                                           np.max(diff_airy))
+    assert np.max(diff_airy) < 0.01, ("Difference between exact Airy pattern and calculated by the numerical integral is bigger than 1%:"
+                                      + str(np.max(diff_airy)))
     # Test normal calculation of several polynomials
-    pols = (ZernPol(osa=10), ZernPol(osa=15)); coeffs = (0.3, -0.27)
-    zpsf = ZernPSF(pols); zpsf.set_physical_props(NA, wavelength, expansion_coeff=coeffs, pixel_physical_size=wavelength / 3.5)
-    zpsf.set_calculation_props(kernel_size=25, n_integration_points_r=200, n_integration_points_phi=180)
-    psf_kernel = zpsf.calculate_psf_kernel()
+    pols = (ZernPol(osa=10), ZernPol(osa=15)); coeffs = (0.08, -0.07)
+    zpsf = ZernPSF(pols); zpsf.set_physical_props(NA, wavelength, expansion_coeff=coeffs, pixel_physical_size=wavelength / 2.8)
+    zpsf.set_calculation_props(kernel_size=23, n_integration_points_r=150, n_integration_points_phi=120)
+    psf_kernel = zpsf.calculate_psf_kernel(); psf_kernel_size = zpsf.kernel_size
+    w_orig_kernel, h_orig_kernel = psf_kernel.shape
     assert np.max(psf_kernel) > 0.5 and np.min(psf_kernel) > -1E-5, "Check calculation of a kernel for several polynomials"
+    # Test cropping of calculated kernel
+    zpsf.crop_kernel()  # cropping the calculated kernel
+    cropped_kernel = zpsf.kernel; cropped_kernel_size = zpsf.kernel_size
+    w_crop_kernel, h_crop_kernel = cropped_kernel.shape
+    assert psf_kernel_size > cropped_kernel_size, (f"Cropped kernel size ({psf_kernel_size}) is equal "
+                                                   + f"or more than original ({cropped_kernel_size})")
+    assert w_orig_kernel > w_crop_kernel and h_orig_kernel > h_crop_kernel, (f"Cropped kernel shape ({cropped_kernel.shape}) is equal "
+                                                                             + f"or more than original({psf_kernel.shape})")
 
 
 def test_zernpsf_usage():
     NA = 0.95; wavelength = 0.55; pixel_size = wavelength / 5.0; ampl = 0.55  # Common physical properties
     zp1 = ZernPol(m=0, n=2)  # defocus
-    test_init = True
+    test_init = True  # flag, if set to False, the test failed (e.g., expected Error not caught)
+    # Test for a wrong ZernPol
     try:
         zpsf = ZernPSF(zernpol=(0, 2)); test_init = False
     except ValueError:
         pass
-    assert test_init, print("ZernPSF wrong initialization passed through")
+    assert test_init, "ZernPSF initialization with wrong ZernPol (as a tuple with orders) passed through"
     zpsf = ZernPSF(zp1)  # proper initialization
+    # Wrong NA provided
     try:
         zpsf.set_physical_props(NA=2.0, wavelength=wavelength, expansion_coeff=ampl, pixel_physical_size=pixel_size)  # wrong NA
         test_init = False
     except ValueError:
         pass
-    assert test_init, print("Wrong NA assigned but no ValueError thrown")
+    assert test_init, "Wrong NA assigned (=2.0) but no ValueError thrown"
+    # Wrong kernel size set
     zpsf.set_physical_props(NA=NA, wavelength=wavelength, expansion_coeff=ampl, pixel_physical_size=pixel_size)  # normal assignment
-    zpsf.set_calculation_props(kernel_size=zpsf.kernel_size, n_integration_points_r=250, n_integration_points_phi=300)  # normal assignment
+    zpsf.set_calculation_props(kernel_size=zpsf.kernel_size, n_integration_points_r=225, n_integration_points_phi=240)  # normal assignment
     try:
         zpsf.set_calculation_props(kernel_size=-1, n_integration_points_r=250, n_integration_points_phi=300)
         test_init = False
     except ValueError:
         pass
-    assert test_init, print("Wrong kernel size assigned but no ValueError thrown")
+    assert test_init, "Wrong kernel size (=-1) assigned but no ValueError thrown"
+    # Several polynomials provided but with single amplitude
+    zpsf = ZernPSF([ZernPol(osa=3), ZernPol(osa=10)]); ampls = (0.21)  # test case - 2 pol-s, 1 ampl.
+    std_na = 0.95; std_wl = 0.52; std_pixel_size = 6.5 / 63  # standard physical parameters
+    try:
+        zpsf.set_physical_props(NA=std_na, wavelength=std_wl, expansion_coeff=ampls, pixel_physical_size=std_pixel_size)
+        test_init = False
+    except ValueError:
+        pass
+    assert test_init, "2 pol-s provided but with only 1 amplitude and not ValueError thrown"
+    # Several polynomials provided but with more amplitudes than polynomials + provided Piston
+    zpsf = ZernPSF((ZernPol(osa=3), ZernPol(osa=0))); ampls = [0.1, -0.2, -1.0]  # test case - 2 pol-s, 3 ampl-s
+    try:
+        zpsf.set_physical_props(NA=std_na, wavelength=std_wl, expansion_coeff=ampls, pixel_physical_size=std_pixel_size)
+        test_init = False
+    except ValueError:
+        pass
+    assert test_init, "2 pol-s provided but with 3 amplitudes and not ValueError thrown"
+    # Single polynomial + amplitude as Sequence provided
+    ampls = [0.75]; ampls = np.asarray(ampls); pols = (ZernPol(noll=14), )  # test case
+    zpsf = ZernPSF(pols); zpsf.set_physical_props(NA=std_na, wavelength=std_wl, expansion_coeff=ampls, pixel_physical_size=std_pixel_size)
 
 
 def test_save_load_zernpsf():
     zp2 = ZernPol(m=1, n=3); zpsf2 = ZernPSF(zp2)  # horizontal coma
     zp3 = ZernPol(osa=21); zpsf3 = ZernPSF(zp3)  # additional classes for testing reading and reassigning values
-    NA = 0.4; wavelength = 0.4; pixel_size = wavelength / 3.2; ampl = 0.16  # Common physical properties
+    NA = 0.4; wavelength = 0.4; pixel_size = wavelength / 3.0; ampl = 0.11  # Common physical properties
     zpsf2.set_physical_props(NA=NA, wavelength=wavelength, expansion_coeff=ampl, pixel_physical_size=pixel_size)
     zpsf2.set_calculation_props(kernel_size=zpsf2.kernel_size, n_integration_points_r=200, n_integration_points_phi=180)
-    zpsf3.set_physical_props(NA=1.0, wavelength=0.6, expansion_coeff=-0.1, pixel_physical_size=0.05)
+    zpsf3.set_physical_props(NA=1.0, wavelength=0.6, expansion_coeff=-0.08, pixel_physical_size=0.06)
     zpsf2.calculate_psf_kernel(normalized=True)  # normal calculation of a kernel
     zpsf2.save_json(overwrite=True)  # save in the standard location (package folder)
     assert Path(zpsf2.json_file_path).is_file(), "File hasn't been saved in the standard location"
@@ -88,7 +120,7 @@ def test_save_load_zernpsf():
 def test_numba_compilation():
     force_get_psf_compilation()  # force compilation of computation methods
     # Test the difference between accelerated and not accelerated calculation methods
-    NA = 0.95; wavelength = 0.55; pixel_size = wavelength / 4.6; ampl = -0.16
+    NA = 0.95; wavelength = 0.55; pixel_size = wavelength / 4.25; ampl = -0.12
     zp6 = ZernPol(m=0, n=2); zpsf6 = ZernPSF(zp6); zpsf7 = ZernPSF(zp6)  # defocus
     zpsf6.set_physical_props(NA=NA, wavelength=wavelength, expansion_coeff=ampl, pixel_physical_size=pixel_size)
     zpsf7.set_physical_props(NA=NA, wavelength=wavelength, expansion_coeff=ampl, pixel_physical_size=pixel_size)